refactor(playback::playlist): move sorting to own module
